{BKR-622} Fix Regular Expression to give correct Host IP Address
Without this patch applied the hostname can sometimes be incorrectly picked up and this
creates problem with all the hostnames getting the same IP addresses on a multi-node setup.

Example Nodeset that can cause this problem:
HOSTS:
  staging:
    roles:
      - sta
      - master
    platform: el-6-x86_64
    box: puppetlabs/centos-6.6-64-puppet
    box_url: https://vagrantcloud.com/puppetlabs/boxes/centos-6.6-64-puppet
    hypervisor: vagrant
    shared_folder: abc
    NetworkSettings:
      IPAddress: 10.255.50.100
  etl:
    roles:
      - etl
    platform: el-6-x86_64
    box: puppetlabs/centos-6.6-64-puppet
    box_url: https://vagrantcloud.com/puppetlabs/boxes/centos-6.6-64-puppet
    hypervisor: vagrant
    NetworkSettings:
      IPAddress: 10.255.50.120
CONFIG:
  type: git
  destroy: no

In the above scenario, the hostname 'etl' is also available when you search 'puppetlabs'.

The patch fixes this by changing the regular expression to look for the string 'etl' (including
the quote marks).